break up tests into smaller test classes
