+++ src/main/kotlin/me/ntrrgc/tsGenerator/VoidType.kt
